requirements.txt
googletrans==3.1.0a0
gTTS==2.2.3
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
pytesseract==0.3.8

